Implement Snackbar notifications for new releases
An OkHttp interceptor has been added to capture the headers notifying of
new versions of the app that have been release on the app stores. These
are broadcasted over the event bus and displayed on
MyCoursesListActivity (the main dashboard Activity) in the form of
Snackbars, that will persist until the user swipes them away. The
maximum limit for the number of lines in the Snackbar messages has also
been raised from 2 to 10, to stop the error messages from getting
truncated.

The Activity is treated as a session, with notifications not being
repeated in a single session, but reset after the session is closed.
This design is so that the notifications are always provided directly by
the server, instead of being persisted on the local device. This will
prevent wrongly sent headers to permanently change the app state to an
incorrect one.

There are three different notifications in descending order of priority.
New notifications will only be shown if they have a higher priority than
any other one that has been shown in the current session.

1) Unsupported app version notifications. This is when we get a 426
(upgrade required) error as a response. This means that the API we're
calling has been removed, and will no longer deliver anything.

2) Deprecation notification. This is when we get notified of a new
version of the app, along with a deadline after which the current
version won't have it's API calls supported any longer. These are
prioritized according to the closeness of their deadlines.

3) New release notification. This is when we get notified of a new
version of the app, without any deadline for support of the current
version.  These are prioritized according to the largest new version
number.

Apache Commons Lang 3 has been added as a new dependency, since it's
FastDateFormat is a very useful improvement over the built-in
SimpleDateFormat, and provides much better support for ISO 8601, which
seems to be the standard used by the dates sent by the server.
Specifically, the last supported date header is being sent with UTC
offsets, which is part of the ISO 8601 standard, and also supported by
FastDateFormat. All SimpleDateFormat usages should be replaced with
FastDateFormat in the future.

Note that we already has Apache Commons Lang 2 imported as a transitive
dependency of Apache Commons IO, which doesn't seem to have a version 3
released, but it's version of FastDateFormat only supports formatting
strings and not parsing them. However, Commons Lang 3 uses a different
package name, so the two versions can coexist without issue.