Remove usages of compile-only Apache Commons Lang
Since commit 5158557b9df55463301a58c3b186716b14cf330a, Apache Commons
Lang 2.4 was being imported as a transitive dependency of the Apache
Velocity Engine library, which itself is a transitive dependency of
RoboBlender, which is an annotation preprocessor that is used by
RoboGuice, which were added in that commit.

Even though this is a compile-only dependency, since the library was
present in the classpath, it began to be used in the actual runtime code
as well, resulting in NoClassDefFoundError being thrown at runtime
instead of a compile-time error. This commit imports the main utility
classes from the latest release version (3.3.4) or the master branch of
the library that were being used in the project, modifying them as
appropriate to make them standalone utilities, linking only to each
other among the imported classes. The release version was used in the
DateUtils class instead of the version on the master branch, because
it's parser used the library's own FastDateParser, which was not
imported, instead of the global SimpleDateFormat.

The StringUtils.equals() utility method was being used by
UserProfileBioModel, but the utility class doesn't seem to be that
useful, and the usage can be replaced by the equivalent method in
ObjectUtils, so it has not been included.

The DateUtils utility class was being used only in the tests, and it was
actually even working fine, but since this is something of an accidental
dependency, it has also been imported over to the test package (which
won't be included in the APK).

FastDateFormat was not included, because it's only used in one place,
which is about to be removed for now anyway.