Support parsing dates according to ISO 8601
Gson 2.5 imported a parser (and formatter) utility class for ISO 8601
compliant date strings from the Jackson library, which is used as a
fallback by it's date type adapters. It doesn't have a dedicated type
adapter for this that would also serialize to this format though. This
commit updates Gson 2.4 to the latest release at 2.7, and adds a type
adapter that uses it's parser to convert between an ISO 8601 compliant
date string and a Date object. The existing custom wrapper for the date
type adapters that parsed a few hardcoded patterns has been removed.

The DateUtil utility class has also been updated to use the Gson parser
and formatter instead of just parsing the standard pattern using
SimpleDateFormat, and formatting more complex incarnations manually. The
getCurrentTimeStamp() method has been changed from using a custom user-
readable format to an ISO 8601 complaint format, and the
getModificationDate() method that was previously used for this purpose
has been replaced with it. The original version was used to log the time
for HTTP requests, but Logcat already displays timestamps along with
each message, so this shouldn't be needed anywhere. The test for it has
also been updated to just verify that it's an ISO 8601 compliant string
(previously it just checked the length), since it's difficult to mock
the current time for testing, and creating an additional method would
probably be overkill. The unused utility method for converting minutes
to hours has been deleted.

The NewVersionBroadcastInterceptor implementation has also been updated
to use DateUtil to parse the date instead of using Apache Commons Lang's
FastDateFormat, which supports parsing the standard ISO 8601 patterns
individually.