Update Snackbar notification strings for updates
The existing update notification strings were too long, and weren't
fitting in the max 2 line limit of the Snackbar message text, so the
maximum limit was manually raised (by finding the text view by ID, since
Snackbar doesn't expose a public API for this). After consultation with
@catong and @lizcohen, shorter strings were suggested for this that fit
within 2 lines on standard devices with standard font size setups.
According to the new suggested strings, the deadline date on deprecation
notifications would not be shown, and it would display the same string
as non-deprecated update notifications. This commit implements this
change, and removes the manual raising of the maximum lines limit on the
Snackbar message text.

The "Upgrade" button text has been changed to "Update" instead, as
suggested by @bguertin. The toast error message string for when the
upgrade button is pressed while no app store or browser is registered on
the app has also been updated to say "Google Play Store" instead of just
"Play Store", as advised by @catong.